Fixed multiple iso mount points for VirtIO drivers
When multiple images are created using the same iso drivers,
the Get-DiskImage command will return the main mount point,
and not all the mountpoints created.

As a consequence, one instance of imaging tools can unmount
the other's virtio ISO, leading to a failed image generation,
as VirtIO drivers cannot be found.