Fixed multiple iso mount points for VirtIO
If multiple images are generated using the same VirtIO ISO,
there is a chance that the imaginfg tool will dismount the VirtIO
ISO at a path in use, which triggers an error to apply the
drivers.

Copying the ISO to a backup file and mounting that ISO will
fix the issue.